MA-99: Moved pan & multiselect logic to MainContentView
